factoring out spark and dask code from utils into their own module
